chore(AnnotationEditor): move the component to components/EditWithPreview
+ extract base component EditTextWithPreview
+ extract EditAnnotationWithPreview
