+++ packages/ui/src/ui/components/EditTextWithPreview/EditTextWithPreview.scss
@@ -1,4 +1,4 @@
-.annotation-editor {
+.edit-text-with-preview {
     display: flex;
     position: relative;
