+++ packages/ui/src/ui/components/EditTextWithPreview/TabbedContent.scss
@@ -1,4 +1,4 @@
-@import '../../../styles/mixins.scss';
+@import '../../styles/mixins.scss';
 
 .yt-tabbed-content {
     display: flex;
