+++ packages/ui/src/ui/pages/navigation/modals/AttributesEditor.scss
@@ -13,39 +13,18 @@
         color: var(--warning-text);
     }
 
-    .df-dialog__tabs-content {
-        overflow: hidden;
-    }
-
     .df-dialog__field-group_type_annotation {
-        flex-grow: 1;
-        overflow: hidden;
-    }
-
-    .df-dialog__field-group_type_annotation {
-        justify-content: flex-start;
-        display: flex;
-
         .df-dialog__field-wrapper_type_annotation {
-            flex-grow: 1;
-            flex-shrink: 1;
-            display: flex;
-            flex-direction: column;
-            width: unset;
-            height: 50vh;
+            padding-top: 0;
         }
 
         .textarea__control {
-            height: 250px;
-        }
-
-        .df-dialog__label {
-            display: none;
+            //height: 250px;
         }
     }
 
     &__full-height {
-        height: 100%;
+        height: 473px;
     }
 
     &__path-list {
